requirements.txt
certifi==2018.4.16
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
scipy==1.1.0
six==1.11.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.23.4
